Refactoring AI file
