+++ capstone_design/fastapi_app/app/open_ai.py
